EES-4921 Refactor out `DataSetVersionsController` in public API
